requirements.txt
paho-mqtt==1.6.1
firebase-admin==6.0.1
Flask==2.2.2
